Slammed out the hardware design for the brains. Needs some cleanup. WIP.
Noticed that it could use a bit of a dramatic redesign with regards to
where the potentiometers are. They're very close to the edge of the
board, which might make them ergonomically hard to use. Given that
there's a lot of extra space I thought I'd stop here and try again in
another branch. Hooray for git